requirements.txt
streamlit>=1.40.0
openai>=1.60.0
python-dotenv>=1.0.1
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
